split instrument(_method)_TestCase.py to two test scripts
